Bugfix: LagQuiz.js printer ut fullstendig quiz i konsollen nå quiz.js: Fikset bug der tidligere spørsmål blir med videre scoreboard.js: Fikset bug der quiz ble slettet selv om du bare observerte, og ikke var med på selve quizzen. Dette krevde også bitteliten endring i quizoversikt.js

+++ src/main/webapp/javascript/scoreboard.js
@@ -95,11 +95,17 @@ function hentSpillere() {
 
 function fjernQuiz() {
     //console.log("FjernQuiz er deaktivert av debug-grunner")
-    $.ajax({
-        url: 'rest/QuizService/quiz/'+quizId+'',
-        type: 'DELETE',
-        success: function(result) {
-            console.log("Fjernet quiz fra indeks "+result);
-        }
-    });
+    if (kallenavn == "" || kallenavn == " " || kallenavn == "undefined")  {
+        console.log("Kallenavn er tomt, ikke slett quiz!")
+    }
+    else  {
+        $.ajax({
+            url: 'rest/QuizService/quiz/'+quizId+'',
+            type: 'DELETE',
+            success: function(result) {
+                console.log("Fjernet quiz fra indeks "+result);
+            }
+        });
+    }
+
 }
